// apps/api/src/schemas/auth.ts
import { z } from 'zod';

export const registerSchema = z.object({
  email: z.string().email(),
  password: z.string().min(8),
  org_name: z.string().min(1).max(100),
  user_name: z.string().min(1).max(100)
});

export const loginSchema = z.object({
  email: z.string().email(),
  password: z.string()
});

export const refreshSchema = z.object({
  refresh_token: z.string()
});